📝 Refactor project structure by creating new module directories and updating imports
