+++ .gitignore
@@ -4,6 +4,6 @@ page_old.py
 tests/super_login_old.py
 *.tmp.py
 *.txt
-super/login_bots/jar.py
-super/login_bots/mwclient.pyyy.py
+super/S_Login/jar.py
+super/S_Login/mwclient.pyyy.py
 wait_time_errors.bash
