+++ mdwiki_page.py
@@ -45,9 +45,9 @@ if "mwclient" not in sys.argv:
     print("sys.argv.append('nomwclient')")
 
 import os
-from newapi.super import bot_api
-from newapi.super import super_page
-from newapi.super import catdepth_new
+from newapi.super.S_API import bot_api
+from newapi.super.S_Page import super_page
+from newapi.super.S_Login import catdepth_new
 
 from newapi import user_account_new
