+++ page.py
@@ -26,9 +26,9 @@ from newapi.page import MainPage, NEW_API
 # ---
 import os
 import sys
-from newapi.super import bot_api
-from newapi.super import super_page
-from newapi.super import catdepth_new
+from newapi.super.S_API import bot_api
+from newapi.super.S_Page import super_page
+from newapi.super.S_Login import catdepth_new
 
 from newapi import useraccount
