+++ super/S_API/bot_api.py
@@ -47,8 +47,8 @@ import sys
 import datetime
 from datetime import timedelta
 from newapi import printe
-from newapi.super.botapi_bots.bot import BOTS_APIS
-from newapi.super.super_login import Login
+from newapi.super.S_API.bot import BOTS_APIS
+from newapi.super.S_Login.super_login import Login
 
 change_codes = {
     "bat_smg": "bat-smg",
