+++ super/S_Category/catdepth_new.py
@@ -10,8 +10,8 @@ import time
 import sys
 import tqdm
 from newapi import printe
-from newapi.super.botapi_bots.bot import BOTS_APIS
-from newapi.super.super_login import Login
+from newapi.super.S_API.bot import BOTS_APIS
+from newapi.super.S_Login.super_login import Login
 
 SITECODE = "en"
 FAMILY = "wikipedia"
