+++ super/S_Page/ar_err.py
@@ -2,19 +2,17 @@
 
 """
 
-from newapi.super.page_bots.ar_err import find_edit_error
-
-if find_edit_error(old, new):
-    return
-
+from newapi.super.S_Page.ar_err import find_edit_error
+if find_edit_error(old, new): return
 """
 
 
 def find_edit_error(old, new):
     # Define the dictionary of conversion phrases
     conversion_phrases = {
-        "#تحويل [[",
+        "#تحويل [["
     }
+
     for phrase in conversion_phrases:
         if phrase in old and phrase not in new:
             print(f"ar_err.py found ({phrase}) in old but not in new. return True")
