+++ super/S_Page/super_page.py
@@ -54,9 +54,9 @@ from newapi import printe
 from newapi import txtlib
 from newapi import botEdit
 
-from newapi.super.page_bots.ar_err import find_edit_error
-from newapi.super.page_bots.bot import APIS
-from newapi.super.super_login import Login
+from newapi.super.S_Page.ar_err import find_edit_error
+from newapi.super.S_Page.bot import APIS
+from newapi.super.S_Login.super_login import Login
 from newapi.except_err import exception_err, warn_err
 
 file_name = os.path.basename(__file__)
@@ -108,7 +108,7 @@ class MainPage(Login, APIS):
         # ---
         """
         Initializes a MainPage instance for interacting with a MediaWiki page.
-        
+
         Sets up page attributes including title, language, family, API endpoint, and metadata fields. Normalizes the language code, loads user tables if available, and logs into the wiki if required.
         """
         self.username = ""
@@ -179,13 +179,13 @@ class MainPage(Login, APIS):
     def ask_put(self, nodiff=False, ASK=False):
         """
         Prompts the user to confirm saving changes to a page, optionally displaying a diff.
-        
+
         If enabled by command-line arguments or parameters, shows the difference between the current and new text, displays summary information, and asks the user to accept or reject the changes. Supports skipping further prompts for subsequent edits.
-        
+
         Args:
             nodiff: If True, skips displaying the diff.
             ASK: If True, forces the prompt regardless of command-line arguments.
-        
+
         Returns:
             True if the user accepts the changes or prompting is not required; False otherwise.
         """
@@ -225,7 +225,7 @@ class MainPage(Login, APIS):
         # ---
         """
         Determines if a proposed edit should be considered erroneous and aborted.
-        
+
         Returns True if the edit is likely to be a false or destructive edit, such as removing over 90% of the page content or matching language-specific error conditions (e.g., for Arabic pages). Returns False if the edit is allowed or if the page is not in the main namespace.
         """
         if self.ns is False or self.ns != 0:
@@ -253,10 +253,10 @@ class MainPage(Login, APIS):
     def import_page(self, family="wikipedia"):
         """
         Imports the page from another wiki family using the MediaWiki API.
-        
+
         Args:
             family: The source wiki family from which to import the page (default is "wikipedia").
-        
+
         Returns:
             The API response data from the import operation.
         """
@@ -280,9 +280,9 @@ class MainPage(Login, APIS):
     def find_create_data(self):
         """
         Retrieves and stores the creation metadata of the page's first revision.
-        
+
         Queries the MediaWiki API for the earliest revision of the page and extracts the creation timestamp, user, and anonymity status. The data is stored in the `create_data` attribute and returned as a dictionary.
-        
+
         Returns:
             dict: A dictionary containing 'timestamp', 'user', and 'anon' keys for the page's creation.
         """
@@ -319,12 +319,12 @@ class MainPage(Login, APIS):
     def get_text(self, redirects=False):
         """
         Retrieves the current wikitext content and metadata for the page.
-        
+
         Fetches the latest revision's wikitext, user, revision ID, timestamp, namespace, Wikibase item, and flagged status. Updates instance attributes with the retrieved data. If the revision is the first (creation), stores creation metadata. Returns the page's wikitext content.
-        
+
         Args:
             redirects: If True, follows redirects to the target page.
-        
+
         Returns:
             The wikitext content of the page.
         """
@@ -407,7 +407,7 @@ class MainPage(Login, APIS):
         # ---
         """
         Fetches and updates comprehensive metadata for the current page from the MediaWiki API.
-        
+
         Retrieves and stores categories (including hidden), language links, templates, backlinks, interwiki links, and page information such as namespace, page ID, length, last revision ID, and last touched timestamp. Updates instance attributes with the fetched data for later access.
         """
         params = {
@@ -719,9 +719,9 @@ class MainPage(Login, APIS):
         # ---
         """
         Returns whether the page is flagged for review or quality control.
-        
+
         If the page text has not been loaded, it is retrieved before checking the flagged status.
-        
+
         Returns:
             bool: True if the page is flagged, False otherwise.
         """
@@ -733,7 +733,7 @@ class MainPage(Login, APIS):
     def get_create_data(self):
         """
         Returns the page creation metadata, fetching it if not already loaded.
-        
+
         The creation metadata includes the timestamp, user, and anonymity status of the first revision.
         """
         if not self.create_data:
@@ -743,7 +743,7 @@ class MainPage(Login, APIS):
     def get_timestamp(self):
         """
         Returns the timestamp of the latest revision of the page.
-        
+
         If the timestamp is not already loaded, retrieves the page content to obtain it.
         """
         if not self.timestamp:
@@ -776,9 +776,9 @@ class MainPage(Login, APIS):
     def save(self, newtext="", summary="", nocreate=1, minor="0", tags="", nodiff=False, ASK=False):
         """
         Saves new text to the page, updating its content and metadata.
-        
+
         Prompts for confirmation and checks for invalid edits before submitting the change. Updates instance attributes with the latest revision and timestamps on success.
-        
+
         Args:
         	newtext: The new wikitext to save to the page.
         	summary: Edit summary for the change.
@@ -787,7 +787,7 @@ class MainPage(Login, APIS):
         	tags: Optional tags to associate with the edit.
         	nodiff: If True, skips showing a diff before saving.
         	ASK: If True, prompts the user for confirmation before saving.
-        
+
         Returns:
         	True if the edit was successful, False otherwise.
         """
@@ -899,15 +899,15 @@ class MainPage(Login, APIS):
         # ---
         """
         Creates a new page with the specified text and summary.
-        
+
         If user confirmation is not suppressed, prompts the user before creating the page. Only succeeds if the page does not already exist. Updates instance attributes with the new page state on success.
-        
+
         Args:
             text: The wikitext content to use for the new page.
             summary: Edit summary for the page creation.
             nodiff: If set, disables showing a diff before confirmation.
             noask: If True, skips the user confirmation prompt.
-        
+
         Returns:
             True if the page was created successfully, False otherwise or if the user aborts.
         """
